Larger GT-Predict data
Use GT-Predict data from xls file instead of the .txt files they used
for decision tree training. This also means we aren't accidentally using
donors as acceptors anymore. Also adding a few files to calculate
conflicts, where we now only have 5 conflicting datapoints with
different reaction bool depending on if you ask GT-Predict or in-house
rate data. We keep both since it's so little data and with random
sampling in randomforest we will get an effect of somewhere in-between.