SMILES input tested
A file with a mix of CIDs and SMILESs as well as a file with only
SMILESs were both tested to work with the chemical feature pipeline.
They were run from the results/*chemicalFeatures/ folder.
Near perfect correlation was also found to previous values curated from
pubchem instead of calculated with RDKit.